+++ crates/miden-testing/src/mock_chain/chain.rs
@@ -1,4 +1,5 @@
 use alloc::{
+    boxed::Box,
     collections::{BTreeMap, BTreeSet},
     vec::Vec,
 };
@@ -14,8 +15,8 @@ use miden_objects::{
     MAX_BATCHES_PER_BLOCK, MAX_OUTPUT_NOTES_PER_BATCH, NoteError, ProposedBatchError,
     ProposedBlockError,
     account::{
-        Account, AccountBuilder, AccountHeader, AccountId, AccountIdAnchor, AccountStorageMode,
-        AccountType, StorageSlot, delta::AccountUpdateDetails,
+        Account, AccountBuilder, AccountId, AccountIdAnchor, AccountStorageMode, AccountType,
+        StorageSlot, delta::AccountUpdateDetails,
     },
     asset::{Asset, TokenSymbol},
     batch::{ProposedBatch, ProvenBatch},
@@ -27,9 +28,9 @@ use miden_objects::{
     note::{Note, NoteHeader, NoteId, NoteInclusionProof, NoteType, Nullifier},
     testing::account_code::DEFAULT_AUTH_SCRIPT,
     transaction::{
-        ExecutedTransaction, ForeignAccountInputs, InputNote, InputNotes,
-        OrderedTransactionHeaders, OutputNote, PartialBlockchain, ProvenTransaction,
-        TransactionHeader, TransactionInputs, TransactionScript,
+        AccountInputs, ExecutedTransaction, InputNote, InputNotes, OrderedTransactionHeaders,
+        OutputNote, PartialBlockchain, ProvenTransaction, TransactionHeader, TransactionInputs,
+        TransactionScript,
     },
 };
 use rand::{Rng, SeedableRng};
@@ -676,7 +677,7 @@ impl MockChain {
     }
 
     /// Gets foreign account inputs to execute FPI transactions.
-    pub fn get_foreign_account_inputs(&self, account_id: AccountId) -> ForeignAccountInputs {
+    pub fn get_foreign_account_inputs(&self, account_id: AccountId) -> AccountInputs {
         let account = self.committed_account(account_id);
 
         let account_witness = self.account_tree().open(account_id);
@@ -691,13 +692,7 @@ impl MockChain {
             }
         }
 
-        ForeignAccountInputs::new(
-            AccountHeader::from(account),
-            account.storage().get_header(),
-            account.code().clone(),
-            account_witness,
-            storage_map_proofs,
-        )
+        AccountInputs::new(account.into(), account_witness)
     }
 
     /// Gets the inputs for a block for the provided batches.
@@ -1459,7 +1454,7 @@ pub enum AccountState {
 pub enum TxContextInput {
     AccountId(AccountId),
     Account(Account),
-    ExecutedTransaction(ExecutedTransaction),
+    ExecutedTransaction(Box<ExecutedTransaction>),
 }
 
 impl From<AccountId> for TxContextInput {
@@ -1476,7 +1471,7 @@ impl From<Account> for TxContextInput {
 
 impl From<ExecutedTransaction> for TxContextInput {
     fn from(tx: ExecutedTransaction) -> Self {
-        Self::ExecutedTransaction(tx)
+        Self::ExecutedTransaction(Box::new(tx))
     }
 }