+++ crates/miden-tx/src/prover/mod.rs
@@ -1,6 +1,6 @@
 #[cfg(feature = "async")]
 use alloc::boxed::Box;
-use alloc::{sync::Arc, vec::Vec};
+use alloc::{collections::BTreeSet, sync::Arc, vec::Vec};
 
 use miden_lib::transaction::TransactionKernel;
 use miden_objects::{
@@ -10,7 +10,7 @@ use miden_objects::{
 };
 pub use miden_prover::ProvingOptions;
 use miden_prover::prove;
-use vm_processor::MemAdviceProvider;
+use vm_processor::{Digest, MemAdviceProvider};
 use winter_maybe_async::*;
 
 use super::{TransactionHost, TransactionProverError};
@@ -91,16 +91,18 @@ impl TransactionProver for LocalTransactionProver {
         // load the store with account/note/tx_script MASTs
         self.mast_store.load_transaction_code(account.code(), input_notes, &tx_args);
 
+        let account_code_commitments: BTreeSet<Digest> = tx_args
+            .foreign_account_inputs()
+            .iter()
+            .map(|acc| acc.code().commitment())
+            .collect();
+
         let mut host: TransactionHost<_> = TransactionHost::new(
             account.into(),
             advice_provider,
             self.mast_store.clone(),
             None,
-            tx_args
-                .foreign_accounts()
-                .iter()
-                .map(|acc| acc.account_code().commitment())
-                .collect(),
+            account_code_commitments,
         )
         .map_err(TransactionProverError::TransactionHostCreationFailed)?;
 
@@ -140,7 +142,8 @@ impl TransactionProver for LocalTransactionProver {
         .add_input_notes(input_notes)
         .add_output_notes(output_notes);
 
-        let builder = match account.is_public() {
+        // If the account is on-chain, add the update details.
+        let builder = match account.is_onchain() {
             true => {
                 let account_update_details = if account.is_new() {
                     let mut account = account.clone();